+++ example/lib/example5_themes_playground/widgets/panels/menu_settings/menu_settings.dart
@@ -43,31 +43,43 @@ class MenuSettings extends StatelessWidget {
         const SizedBox(height: 8),
         const ListTile(
           title: Text('Menus'),
-          subtitle: Text('The theming below only concerns the opened menu.'),
+          subtitle: Text('The menu theming properties are shared '
+              'in FCS across multiple component sub-themes to ensure a '
+              'consistent style on all menu components whenever possible.'),
         ),
         ListTile(
           enabled: controller.useSubThemes && controller.useFlexColorScheme,
-          title: const Text('Menu border radius, follows global to max 10'),
-          subtitle: Slider(
-            min: -1,
-            max: 12,
-            divisions: 13,
-            label: controller.useSubThemes && controller.useFlexColorScheme
-                ? controller.popupMenuBorderRadius == null ||
-                        (controller.popupMenuBorderRadius ?? -1) < 0
-                    ? popupMenuDefaultRadiusLabel
-                    : (controller.popupMenuBorderRadius?.toStringAsFixed(0) ??
-                        '')
-                : 'default 4',
-            value: controller.useSubThemes && controller.useFlexColorScheme
-                ? controller.popupMenuBorderRadius ?? -1
-                : -1,
-            onChanged: controller.useSubThemes && controller.useFlexColorScheme
-                ? (double value) {
-                    controller.setPopupMenuBorderRadius(
-                        value < 0 ? null : value.roundToDouble());
-                  }
-                : null,
+          title: const Text('Menu container border radius'),
+          subtitle: Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: <Widget>[
+              const Text('Respects global setting up to max 10. If higher '
+                  'than 10 is used, the selection indicator on '
+                  'PopupMenuButton overflows visibly in the rounded corners.'),
+              Slider(
+                min: -1,
+                max: 12,
+                divisions: 13,
+                label: controller.useSubThemes && controller.useFlexColorScheme
+                    ? controller.popupMenuBorderRadius == null ||
+                            (controller.popupMenuBorderRadius ?? -1) < 0
+                        ? popupMenuDefaultRadiusLabel
+                        : (controller.popupMenuBorderRadius
+                                ?.toStringAsFixed(0) ??
+                            '')
+                    : 'default 4',
+                value: controller.useSubThemes && controller.useFlexColorScheme
+                    ? controller.popupMenuBorderRadius ?? -1
+                    : -1,
+                onChanged:
+                    controller.useSubThemes && controller.useFlexColorScheme
+                        ? (double value) {
+                            controller.setPopupMenuBorderRadius(
+                                value < 0 ? null : value.roundToDouble());
+                          }
+                        : null,
+              ),
+            ],
           ),
           trailing: Padding(
             padding: const EdgeInsetsDirectional.only(end: 12),
@@ -76,7 +88,7 @@ class MenuSettings extends StatelessWidget {
               children: <Widget>[
                 Text(
                   'RADIUS',
-                  style: Theme.of(context).textTheme.bodySmall,
+                  style: theme.textTheme.bodySmall,
                 ),
                 Text(
                   controller.useSubThemes && controller.useFlexColorScheme
@@ -87,9 +99,7 @@ class MenuSettings extends StatelessWidget {
                                   ?.toStringAsFixed(0) ??
                               '')
                       : 'default 4',
-                  style: Theme.of(context)
-                      .textTheme
-                      .bodySmall!
+                  style: theme.textTheme.bodySmall!
                       .copyWith(fontWeight: FontWeight.bold),
                 ),
               ],
@@ -98,28 +108,38 @@ class MenuSettings extends StatelessWidget {
         ),
         ListTile(
           enabled: controller.useSubThemes && controller.useFlexColorScheme,
-          title: const Text('Menu elevation'),
-          subtitle: Slider(
-            min: -1,
-            max: 20,
-            divisions: 21,
-            label: controller.useSubThemes && controller.useFlexColorScheme
-                ? controller.popupMenuElevation == null ||
-                        (controller.popupMenuElevation ?? -1) < 0
-                    ? popupMenuElevationDefaultLabel
-                    : (controller.popupMenuElevation?.toStringAsFixed(0) ?? '')
-                : useMaterial3
-                    ? 'default 3'
-                    : 'default 8',
-            value: controller.useSubThemes && controller.useFlexColorScheme
-                ? controller.popupMenuElevation ?? -1
-                : -1,
-            onChanged: controller.useSubThemes && controller.useFlexColorScheme
-                ? (double value) {
-                    controller.setPopupMenuElevation(
-                        value < 0 ? null : value.roundToDouble());
-                  }
-                : null,
+          title: const Text('Menu container elevation'),
+          subtitle: Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: <Widget>[
+              const Text('Menu containers still have a drop shadow by default '
+                  'in M3. They also use surface elevation tint, which is '
+                  'useful in dark mode.'),
+              Slider(
+                min: -1,
+                max: 20,
+                divisions: 21,
+                label: controller.useSubThemes && controller.useFlexColorScheme
+                    ? controller.popupMenuElevation == null ||
+                            (controller.popupMenuElevation ?? -1) < 0
+                        ? popupMenuElevationDefaultLabel
+                        : (controller.popupMenuElevation?.toStringAsFixed(0) ??
+                            '')
+                    : useMaterial3
+                        ? 'default 3'
+                        : 'default 8',
+                value: controller.useSubThemes && controller.useFlexColorScheme
+                    ? controller.popupMenuElevation ?? -1
+                    : -1,
+                onChanged:
+                    controller.useSubThemes && controller.useFlexColorScheme
+                        ? (double value) {
+                            controller.setPopupMenuElevation(
+                                value < 0 ? null : value.roundToDouble());
+                          }
+                        : null,
+              ),
+            ],
           ),
           trailing: Padding(
             padding: const EdgeInsetsDirectional.only(end: 12),
@@ -152,8 +172,11 @@ class MenuSettings extends StatelessWidget {
           title: const Text('Menu background color'),
           labelForDefault: 'default (surface)',
           subtitle: const Text(
-            'In current version of FCS, only colors that work '
-            'with the contrasting color for surface are usable options.',
+            'Correct contrast pair color will be used as foreground color '
+            'automatically for menu container content in DropDownMenu, '
+            'MenuBar and MenuAnchor. PopupMenuButton currently '
+            'only supports onSurface as content color. The cause is being '
+            'investigated, potential Flutter issue.',
           ),
           index: controller.popupMenuSchemeColor?.index ?? -1,
           onChanged: controller.useSubThemes && controller.useFlexColorScheme
@@ -169,7 +192,7 @@ class MenuSettings extends StatelessWidget {
         ),
         ListTile(
           enabled: controller.useSubThemes && controller.useFlexColorScheme,
-          title: const Text('Menu background opacity'),
+          title: const Text('Menu container background opacity'),
           subtitle: Slider(
             max: 100,
             divisions: 100,
@@ -188,25 +211,27 @@ class MenuSettings extends StatelessWidget {
               children: <Widget>[
                 Text(
                   'OPACITY',
-                  style: Theme.of(context).textTheme.bodySmall,
+                  style: theme.textTheme.bodySmall,
                 ),
                 Text(
                   // ignore: lines_longer_than_80_chars
                   '${(popupOpacity * 100).toStringAsFixed(0)} %',
-                  style: Theme.of(context)
-                      .textTheme
-                      .bodySmall!
+                  style: theme.textTheme.bodySmall!
                       .copyWith(fontWeight: FontWeight.bold),
                 ),
               ],
             ),
           ),
         ),
-        const Divider(),
+        const Divider(height: 32),
         const ListTile(
           title: Text('PopupMenuButton'),
           subtitle: Text('The PopupMenuButton can be used on any kind of '
-              'widget, below a custom TextButton is being used.'),
+              'widget, below a custom TextButton is being used. The '
+              'PopupMenuButton has a different implementation in Flutter '
+              'than the newer M3 menus further below. The properties '
+              'originally used to only theme it, are now also used '
+              'to style equivalent properties in the newer menus.'),
         ),
         Padding(
           padding: const EdgeInsets.symmetric(horizontal: 16),
@@ -222,25 +247,143 @@ class MenuSettings extends StatelessWidget {
             ],
           ),
         ),
-        const SizedBox(height: 16),
+        const Divider(height: 32),
+        // const SizedBox(height: 16),
         const ListTile(
           title: Text('DropdownMenu'),
-          subtitle: Text('New component in M3. Text input area uses the active '
-              'InputDecorationTheme in FCS. The DropdownMenu will get more '
-              'styling options in FCS later versions, e.g. tuning the default '
-              'selection highlight style.'),
+          subtitle: Text('Text input area uses the themed '
+              'InputDecoration settings. Other theming properties are '
+              'shared with the PopupMenuButton.'),
         ),
         const Padding(
           padding: EdgeInsets.symmetric(horizontal: 16),
           child: DropDownMenuShowcase(),
         ),
         const SizedBox(height: 16),
+        const Divider(),
         const ListTile(
           title: Text('MenuBar'),
-          subtitle: Text('New component in M3. MenuBar is made up of '
-              'SubMenuButton with MenuItemButton. The MenuBar will get more '
-              'styling options in FCS later versions.'),
+          subtitle: Text('THe MenuBar is made up of SubMenuButtons that have '
+              'MenuItemButtons. You can construct arbitrary deep nested '
+              'sub-menus and items can have keyboard shortcuts.'),
         ),
+        ListTile(
+          enabled: controller.useSubThemes && controller.useFlexColorScheme,
+          title: const Text('MenuBar border radius'),
+          subtitle: Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: <Widget>[
+              const Text('For a nice edge-to-edge look, consider using 0 dp. '
+                  'The M3 guide depicts MenuBar as edge-to-edge with no '
+                  'corner rounding, but Flutter M3 defaults to 4 dp.'),
+              Slider(
+                min: -1,
+                max: 30,
+                divisions: 31,
+                label: controller.useSubThemes && controller.useFlexColorScheme
+                    ? controller.menuBarRadius == null ||
+                            (controller.menuBarRadius ?? -1) < 0
+                        ? 'default 4'
+                        : (controller.menuBarRadius?.toStringAsFixed(0) ?? '')
+                    : 'default 4',
+                value: controller.useSubThemes && controller.useFlexColorScheme
+                    ? controller.menuBarRadius ?? -1
+                    : -1,
+                onChanged:
+                    controller.useSubThemes && controller.useFlexColorScheme
+                        ? (double value) {
+                            controller.setMenuBarRadius(
+                                value < 0 ? null : value.roundToDouble());
+                          }
+                        : null,
+              ),
+            ],
+          ),
+          trailing: Padding(
+            padding: const EdgeInsetsDirectional.only(end: 12),
+            child: Column(
+              crossAxisAlignment: CrossAxisAlignment.center,
+              children: <Widget>[
+                Text(
+                  'RADIUS',
+                  style: theme.textTheme.bodySmall,
+                ),
+                Text(
+                  controller.useSubThemes && controller.useFlexColorScheme
+                      ? controller.menuBarRadius == null ||
+                              (controller.menuBarRadius ?? -1) < 0
+                          ? popupMenuDefaultRadiusLabel
+                          : (controller.menuBarRadius?.toStringAsFixed(0) ?? '')
+                      : 'default 4',
+                  style: theme.textTheme.bodySmall!
+                      .copyWith(fontWeight: FontWeight.bold),
+                ),
+              ],
+            ),
+          ),
+        ),
+        ListTile(
+          enabled: controller.useSubThemes && controller.useFlexColorScheme,
+          title: const Text('MenuBar elevation'),
+          subtitle: Slider(
+            min: -1,
+            max: 30,
+            divisions: 31,
+            label: controller.useSubThemes && controller.useFlexColorScheme
+                ? controller.menuBarElevation == null ||
+                        (controller.menuBarElevation ?? -1) < 0
+                    ? 'default 3'
+                    : (controller.menuBarElevation?.toStringAsFixed(0) ?? '')
+                : 'default 3',
+            value: controller.useSubThemes && controller.useFlexColorScheme
+                ? controller.menuBarElevation ?? -1
+                : -1,
+            onChanged: controller.useSubThemes && controller.useFlexColorScheme
+                ? (double value) {
+                    controller.setMenuBarElevation(
+                        value < 0 ? null : value.roundToDouble());
+                  }
+                : null,
+          ),
+          trailing: Padding(
+            padding: const EdgeInsetsDirectional.only(end: 12),
+            child: Column(
+              crossAxisAlignment: CrossAxisAlignment.center,
+              children: <Widget>[
+                Text(
+                  'ELEV',
+                  style: theme.textTheme.bodySmall,
+                ),
+                Text(
+                  controller.useSubThemes && controller.useFlexColorScheme
+                      ? controller.menuBarElevation == null ||
+                              (controller.menuBarElevation ?? -1) < 0
+                          ? 'default 3'
+                          : (controller.menuBarElevation?.toStringAsFixed(0) ??
+                              '')
+                      : 'default 3',
+                  style: theme.textTheme.bodySmall!
+                      .copyWith(fontWeight: FontWeight.bold),
+                ),
+              ],
+            ),
+          ),
+        ),
+        SwitchListTile(
+            title: const Text('Remove elevation shadow'),
+            subtitle: const Text('The M3 guide depicts MenuBars with no '
+                'shadow and optional elevation with tint. Flutter defaults '
+                'has shadow in M3. To be able to use elevation with only '
+                'elevation tint in M3, turn ON to remove the shadow.'),
+            value: controller.menuBarShadowColor == Colors.transparent,
+            onChanged: (bool removeShadow) {
+              if (removeShadow) {
+                controller.setMenuBarShadowColor(Colors.transparent);
+              } else {
+                controller.setMenuBarShadowColor(null);
+              }
+            }),
+        const SizedBox(height: 8),
         const Padding(
           padding: EdgeInsets.symmetric(horizontal: 16),
           child: MenuBarShowcase(explainUsage: false),
@@ -248,12 +391,10 @@ class MenuSettings extends StatelessWidget {
         const SizedBox(height: 16),
         const ListTile(
           title: Text('MenuAnchor'),
-          subtitle: Text('New component in M3. The MenuAnchor can be used to '
-              'attach a menu to any widget. It uses the same building blocks '
-              'as the MenuBar, SubMenuButton with MenuItemButton. It can thus '
-              'have sub-menus and keyboard shortcuts, like the MenuBar. '
-              'The MenuAnchor will get more '
-              'styling options in FCS later versions.'),
+          subtitle: Text('The MenuAnchor can be used to attach a menu to any '
+              'widget. It is based on same building blocks as the MenuBar by '
+              'using SubMenuButton with MenuItemButton. It can have sub-menus '
+              'and keyboard shortcuts, like the MenuBar.'),
         ),
         const Padding(
           padding: EdgeInsets.symmetric(horizontal: 16),